Move `DIBuilderBox` out of `ffi.rs`
