+++ compiler/rustc_codegen_llvm/src/debuginfo/di_builder.rs
@@ -1,8 +1,37 @@
+use std::ptr;
+
 use libc::c_uint;
 use rustc_abi::Align;
 
 use crate::llvm::debuginfo::DIBuilder;
-use crate::llvm::{self, ToLlvmBool};
+use crate::llvm::{self, Module, ToLlvmBool};
+
+/// Owning pointer to a `DIBuilder<'ll>` that will dispose of the builder
+/// when dropped. Use `.as_ref()` to get the underlying `&DIBuilder`
+/// needed for debuginfo FFI calls.
+pub(crate) struct DIBuilderBox<'ll> {
+    raw: ptr::NonNull<DIBuilder<'ll>>,
+}
+
+impl<'ll> DIBuilderBox<'ll> {
+    pub(crate) fn new(llmod: &'ll Module) -> Self {
+        let raw = unsafe { llvm::LLVMCreateDIBuilder(llmod) };
+        let raw = ptr::NonNull::new(raw).unwrap();
+        Self { raw }
+    }
+
+    pub(crate) fn as_ref(&self) -> &DIBuilder<'ll> {
+        // SAFETY: This is an owning pointer, so `&DIBuilder` is valid
+        // for as long as `&self` is.
+        unsafe { self.raw.as_ref() }
+    }
+}
+
+impl<'ll> Drop for DIBuilderBox<'ll> {
+    fn drop(&mut self) {
+        unsafe { llvm::LLVMDisposeDIBuilder(self.raw) };
+    }
+}
 
 /// Extension trait for defining safe wrappers and helper methods on
 /// `&DIBuilder<'ll>`, without requiring it to be defined in the same crate.
